修正DSD 移動撥放位置錯誤

+++ src/xamp_output_device/include/output_device/asiodevice.h
@@ -91,7 +91,7 @@ private:
 	mutable double volume_;
 	int32_t buffer_size_;
 	int32_t buffer_bytes_;
-	std::atomic<int32_t> played_bytes_;
+	std::atomic<int64_t> played_bytes_;
 	std::string device_id_;
 	mutable std::mutex mutex_;
 	std::condition_variable condition_;
